// src/types/product.ts
export interface TProductDetails {
  id: string;
  name: string;
  description: string;
  category: string;
  supplierId: string;
  unitPrice: string;
  quantity: string;
  imageUrlList: string[];
  imageUrl: any[];
  createdAt: any;
  updatedAt: any;
  productCategory: any;
  available: boolean;
}